Module gfx_memory::allocator::linear

Structs

Line
If freed >= allocated it is safe to free the line.

LinearAllocator
The LinearAllocator is best suited for short-lived allocations. The allocation strategy has minimal overhead and the implementation is fast. But holding a single block will completely stop memory recycling.

LinearBlock
Memory block allocated from LinearAllocator.

LinearConfig
Config for LinearAllocator. Refer to documentation on LinearAllocator to better understand what the configuration options mean.

Type Definitions

LineCount