Struct gfx_memory::allocator::linear::LinearAllocator

pub struct LinearAllocator<B: Backend> {
    memory_type: MemoryTypeId,
    memory_properties: Properties,
    line_size: Size,
    finished_lines_count: u32,
    lines: VecDeque<Line<B>>,
    non_coherent_atom_size: Option<AtomSize>,
    unused_lines: Vec<Line<B>>,
}

The LinearAllocator is best suited for short-lived allocations. The allocation strategy has minimal overhead and the implementation is fast. But holding a single block will completely stop memory recycling.

The linear allocator will internally create multiple lines. Each line is a gfx_hal::Backend::Memory from which multiple LinearBlocks are linearly allocated.

A new line is created if there is insufficient space to create a LinearBlock from the current line. New lines are created from scratch or taken from a pool of previously used lines. When lines have no allocated LinearBlocks remaining they are moved to a pool to be reused. This pool of unused lines is freed only when LinearAllocator::clear is called.

Fields
memory_type: MemoryTypeIdmemory_properties: Propertiesline_size: Sizefinished_lines_count: u32lines: VecDeque<Line<B>>non_coherent_atom_size: Option<AtomSize>unused_lines: Vec<Line<B>>

Previously used lines that have been replaced, kept around to use next time a new line is needed.

Implementations

impl<B: Backend> LinearAllocator<B>

pub fn new(
memory_type: MemoryTypeId,
memory_properties: Properties,
config: LinearConfig,
non_coherent_atom_size: Size
) -> Self

Create new LinearAllocator for memory_type with memory_properties specified, with config.

pub fn max_allocation(&self) -> Size

Maximum allocation size.

fn cleanup(&mut self, device: &B::Device, free_memory: bool) -> Size

pub fn clear(&mut self, device: &B::Device) -> Size

Perform full cleanup of the allocated memory.