Struct gfx_memory::Heaps

pub struct Heaps<B: Backend> { /* fields omitted */ }

Heaps available on particular physical device.

Implementations

impl<B: Backend> Heaps<B>

pub unsafe fn new(
hal_memory_properties: &MemoryProperties,
config_general: GeneralConfig,
config_linear: LinearConfig,
non_coherent_atom_size: Size
) -> Self

Initialize the new Heaps object.

Safety

All later operations assume the device is not lost.

pub fn allocate(
&mut self,
device: &B::Device,
requirements: &Requirements,
usage: MemoryUsage,
kind: Kind
) -> Result<MemoryBlock<B>, HeapsError>

Allocate memory block give the requirements from gfx-hal. for intended usage, using the kind of allocator.

pub fn free(&mut self, device: &B::Device, block: MemoryBlock<B>)

Free memory block.

Memory block must be allocated from this heap.

pub fn clear(&mut self, device: &B::Device)

Clear allocators. Call this before dropping an instance of Heaps or if you are low on memory.

Internally calls the clear methods on all internal LinearAllocator and GeneralAllocator instances.

pub fn utilization(&self) -> TotalMemoryUtilization

Get memory utilization.